requirements.txt
streamlit
seleniumbase
webdriver-manager
pymysql
SQLAlchemy